+++ docs/conf.py
@@ -212,6 +212,9 @@ def _generate_example_images(app):
     QRCodeEncoder(QRCodeData("親切にしろ", encoding="shift_jis")).save_svg(
         str(out / "qrcode-shift-jis.svg")
     )
+    QRCodeEncoder(
+        QRCodeData.wifi_network(ssid="DoubleDeuceGuest", password="PainDontHurt")
+    ).save_svg(str(out / "qrcode-wifi.svg"))
 
     PDF417Encoder("WDBCA45D2HA327260").save(str(out / "pdf417-example.png"))
     PDF417Encoder("WDBCA45D2HA327260").save_svg(str(out / "pdf417-example.svg"))